requirements.txt
mosspy
pyyaml
python-dotenv
